+++ examples/syscall/module/syscall.go
@@ -19,6 +19,6 @@ func init() {
 }
 
 func messageBox(hWnd uintptr, lpText *uint16, lpCaption *uint16, uType uint32) int32 {
-	r1, _, _ := procMessageBoxW.Call(uintptr(hWnd), uintptr(unsafe.Pointer(lpText)), uintptr(unsafe.Pointer(lpCaption)), uintptr(uType))
+	r1, _, _ := procMessageBoxW.Call(hWnd, uintptr(unsafe.Pointer(lpText)), uintptr(unsafe.Pointer(lpCaption)), uintptr(uType))
 	return int32(r1)
 }
